change projectId from int to long

+++ src/main/java/org/generationcp/commons/hibernate/DefaultManagerFactoryProvider.java
@@ -302,7 +302,7 @@ public class DefaultManagerFactoryProvider implements ManagerFactoryProvider, Ht
         centralSessionProviders.clear();
     }
 
-    public void removeProjectFromLocalSession(int projectId) {
+    public void removeProjectFromLocalSession(long projectId) {
     	if(localSessionFactories!=null && localSessionFactories.containsKey(projectId)) {
     		localSessionFactories.get(projectId).close();
     		localSessionFactories.remove(projectId);
